Add files via upload
Create a batch file named IFLOOP that uses an IF statement and a Counter to control the number of
times a loop is executed. The number of times the loop must be executed depends on the current month.
If the current month is June, since this is the 6th month in the year, the loop must execute 6 times. If the
month is December, the loop must execute 12 times, etc. The value of the current month must be
extracted from the system date. Each iteration of the loop must set the month’s name in a variable
named monthname and must display the value of the counter as well as the value of the monthname
variable. For example, if the current month is May, the loop’s output would look like this:
1 January
2 February
3 March
4 April
5 May
The batch file must work for all months.
Simple IF statements can be used. If you use nested IF statements you will need to address the issue
of delayed expansion.